django/requirements.txt
Django==2.0.4
django-cors-headers==2.2.0
django-modeladmin-reorder==0.3.1
django-rest-framework==0.1.0
# django-versatileimagefield==1.9
djangorestframework-jwt==1.11.0
djangorestframework==3.8.2
gunicorn==19.7.1
psycopg2==2.8.6
tweepy==3.7.0
urllib3==1.24.1
